Litt større omskriving/forenkling av modeller. Henter du et element er det no en NodeDTO, mens barn hentes som NodeChildDTO. Funksjonalitet ellers skal være uendra.
