+++ src/main/java/no/ndla/taxonomy/service/dtos/NodeWithParents.java
@@ -10,31 +10,34 @@ package no.ndla.taxonomy.service.dtos;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
 import no.ndla.taxonomy.domain.Node;
+import no.ndla.taxonomy.domain.exceptions.NotFoundException;
 
 import java.util.HashSet;
 import java.util.Set;
 
 @Schema(name = "ResourceWithParents")
-public class ResourceWithParentsDTO extends ResourceDTO {
+public class NodeWithParents extends NodeDTO {
     @JsonProperty
     @Schema(description = "Parent topology nodes and whether or not connection type is primary", example = "["
             + "{\"id\": \"urn:topic:1:181900\"," + "\"name\": \"I dyrehagen\","
             + "\"contentUri\": \"urn:article:6662\"," + "\"path\": \"/subject:2/topic:1:181900\","
             + "\"primary\": \"true\"}]")
-    private Set<NodeWithResourceConnectionDTO> parents = new HashSet<>();
+    private final Set<NodeChildDTO> parents = new HashSet<>();
 
-    public ResourceWithParentsDTO() {
+    public NodeWithParents() {
+
     }
 
-    public ResourceWithParentsDTO(Node resource, String languageCode) {
-        super(resource, languageCode);
+    public NodeWithParents(Node node, String languageCode) {
+        super(node, languageCode);
 
-        resource.getParentNodeConnections().stream()
-                .map(nodeResource -> new NodeWithResourceConnectionDTO(nodeResource, languageCode))
-                .forEach(parents::add);
+        node.getParentNodeConnections().stream().map(nodeResource -> {
+            Node parent = nodeResource.getParent().orElseThrow(() -> new NotFoundException("Parent not found"));
+            return new NodeChildDTO(parent, nodeResource, languageCode);
+        }).forEach(parents::add);
     }
 
-    public Set<NodeWithResourceConnectionDTO> getParents() {
+    public Set<NodeChildDTO> getParents() {
         return parents;
     }
