+++ src/test/java/no/ndla/taxonomy/rest/v1/NodeTranslationsTest.java
@@ -10,8 +10,6 @@ package no.ndla.taxonomy.rest.v1;
 import no.ndla.taxonomy.domain.Node;
 import no.ndla.taxonomy.domain.NodeType;
 import no.ndla.taxonomy.service.dtos.NodeChildDTO;
-import no.ndla.taxonomy.service.dtos.NodeDTO;
-import no.ndla.taxonomy.service.dtos.ResourceChildDTO;
 import org.junit.jupiter.api.Test;
 
 import java.net.URI;
@@ -158,7 +156,7 @@ public class NodeTranslationsTest extends RestTest {
                 .getPublicId();
 
         var response = testUtils.getResource("/v1/nodes/" + a + "/resources?recursive=true&language=nb");
-        var result = testUtils.getObject(ResourceChildDTO[].class, response);
+        var result = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals(2, result.length);
         assertAnyTrue(result, r -> "Introduksjon til calculus".equals(r.getName()));
@@ -178,7 +176,7 @@ public class NodeTranslationsTest extends RestTest {
                 .child(NodeType.TOPIC, st -> st.name("subtopic").resource(r -> r.name("subtopic resource")))));
 
         var response = testUtils.getResource("/v1/nodes/urn:topic:1/resources?language=nb");
-        var result = testUtils.getObject(ResourceChildDTO[].class, response);
+        var result = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals(2, result.length);
         assertAnyTrue(result, r -> "ressurs 1".equals(r.getName()));
@@ -186,11 +184,11 @@ public class NodeTranslationsTest extends RestTest {
         assertAllTrue(result, r -> r.getResourceTypes().iterator().next().getName().equals("Artikkel"));
     }
 
-    private NodeDTO getNode(URI id, String language) throws Exception {
+    private NodeChildDTO getNode(URI id, String language) throws Exception {
         String path = "/v1/nodes/" + id;
         if (isNotEmpty(language))
             path = path + "?language=" + language;
-        return testUtils.getObject(NodeDTO.class, testUtils.getResource(path));
+        return testUtils.getObject(NodeChildDTO.class, testUtils.getResource(path));
     }
 
 }
