+++ src/test/java/no/ndla/taxonomy/rest/v1/QueryTest.java
@@ -10,7 +10,7 @@ package no.ndla.taxonomy.rest.v1;
 import no.ndla.taxonomy.domain.NodeType;
 import no.ndla.taxonomy.rest.v1.dtos.nodes.searchapi.LanguageField;
 import no.ndla.taxonomy.rest.v1.dtos.nodes.searchapi.TaxonomyContextDTO;
-import no.ndla.taxonomy.service.dtos.NodeDTO;
+import no.ndla.taxonomy.service.dtos.NodeChildDTO;
 import org.junit.jupiter.api.Test;
 import org.springframework.mock.web.MockHttpServletResponse;
 
@@ -28,7 +28,7 @@ public class QueryTest extends RestTest {
                 .resourceType(rt -> rt.name("Subject material")));
 
         MockHttpServletResponse response = testUtils.getResource("/v1/queries/resources?contentURI=urn:article:345");
-        var resources = testUtils.getObject(NodeDTO[].class, response);
+        var resources = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals(1, resources.length);
         assertEquals(resources[0].getResourceTypes().size(), 1);
@@ -37,7 +37,7 @@ public class QueryTest extends RestTest {
     @Test
     public void no_resources_matching_contentURI() throws Exception {
         MockHttpServletResponse response = testUtils.getResource("/v1/queries/resources?contentURI=urn:article:345");
-        var resources = testUtils.getObject(NodeDTO[].class, response);
+        var resources = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals(0, resources.length);
     }
@@ -50,7 +50,7 @@ public class QueryTest extends RestTest {
         builder.node(NodeType.RESOURCE, r -> r.publicId("urn:resource:2").contentUri("urn:article:3"));
 
         MockHttpServletResponse response = testUtils.getResource("/v1/queries/resources?contentURI=urn:article:345");
-        var resources = testUtils.getObject(NodeDTO[].class, response);
+        var resources = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals(1, resources.length);
         assertAnyTrue(resources, r -> "urn:resource:1".equals(r.getId().toString()));
@@ -62,7 +62,7 @@ public class QueryTest extends RestTest {
                 .resourceType(rt -> rt.name("Subject material")).resourceType(rt -> rt.name("Learning path")));
 
         MockHttpServletResponse response = testUtils.getResource("/v1/queries/resources?contentURI=urn:article:345");
-        var resources = testUtils.getObject(NodeDTO[].class, response);
+        var resources = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals(1, resources.length);
         assertEquals(2, resources[0].getResourceTypes().size());
@@ -76,7 +76,7 @@ public class QueryTest extends RestTest {
 
         MockHttpServletResponse response = testUtils
                 .getResource("/v1/queries/resources?contentURI=urn:article:345&language=nb");
-        var resources = testUtils.getObject(NodeDTO[].class, response);
+        var resources = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals(1, resources.length);
         assertEquals(resources[0].getResourceTypes().size(), 1);
@@ -91,7 +91,7 @@ public class QueryTest extends RestTest {
         builder.node(NodeType.TOPIC, r -> r.publicId("urn:topic:2").contentUri("urn:article:345"));
 
         MockHttpServletResponse response = testUtils.getResource("/v1/queries/topics?contentURI=urn:article:345");
-        var resources = testUtils.getObject(NodeDTO[].class, response);
+        var resources = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals(1, resources.length);
         assertAnyTrue(resources, r -> "urn:topic:2".equals(r.getId().toString()));
@@ -104,7 +104,7 @@ public class QueryTest extends RestTest {
         builder.node(NodeType.TOPIC, r -> r.publicId("urn:topic:3").contentUri("urn:article:345"));
 
         MockHttpServletResponse response = testUtils.getResource("/v1/queries/topics?contentURI=urn:article:345");
-        var resources = testUtils.getObject(NodeDTO[].class, response);
+        var resources = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals(2, resources.length);
         assertAnyTrue(resources, r -> "urn:topic:2".equals(r.getId().toString()));
@@ -114,7 +114,7 @@ public class QueryTest extends RestTest {
     @Test
     public void no_topics_matching_contentURI() throws Exception {
         MockHttpServletResponse response = testUtils.getResource("/v1/queries/topics?contentURI=urn:article:345");
-        NodeDTO[] resources = testUtils.getObject(NodeDTO[].class, response);
+        NodeChildDTO[] resources = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals(0, resources.length);
     }
@@ -126,7 +126,7 @@ public class QueryTest extends RestTest {
 
         MockHttpServletResponse response = testUtils
                 .getResource("/v1/queries/topics?contentURI=urn:article:345&language=nb");
-        var resources = testUtils.getObject(NodeDTO[].class, response);
+        var resources = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals(1, resources.length);
         assertEquals("Emne", resources[0].getName());
