+++ src/test/java/no/ndla/taxonomy/rest/v1/ResourcesTest.java
@@ -13,7 +13,10 @@ import no.ndla.taxonomy.domain.Node;
 import no.ndla.taxonomy.domain.NodeType;
 import no.ndla.taxonomy.domain.ResourceType;
 import no.ndla.taxonomy.rest.v1.commands.ResourceCommand;
-import no.ndla.taxonomy.service.dtos.*;
+import no.ndla.taxonomy.service.dtos.NodeChildDTO;
+import no.ndla.taxonomy.service.dtos.NodeDTO;
+import no.ndla.taxonomy.service.dtos.NodeWithParents;
+import no.ndla.taxonomy.service.dtos.ResourceTypeWithConnectionDTO;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -43,7 +46,7 @@ public class ResourcesTest extends RestTest {
                         .name("introduction to trigonometry").contentUri("urn:article:1").publicId("urn:resource:1"))));
 
         final var response = testUtils.getResource("/v1/resources/urn:resource:1");
-        final var resource = testUtils.getObject(ResourceDTO.class, response);
+        final var resource = testUtils.getObject(NodeDTO.class, response);
 
         assertEquals("introduction to trigonometry", resource.getName());
         assertEquals("urn:article:1", resource.getContentUri().toString());
@@ -61,7 +64,7 @@ public class ResourcesTest extends RestTest {
                 t -> t.publicId("urn:topic:2").resource("resource", true)));
 
         var response = testUtils.getResource("/v1/resources/urn:resource:1");
-        final var resource = testUtils.getObject(ResourceDTO.class, response);
+        final var resource = testUtils.getObject(NodeDTO.class, response);
 
         assertEquals("/subject:2/topic:2/resource:1", resource.getPath());
     }
@@ -71,7 +74,7 @@ public class ResourcesTest extends RestTest {
         builder.node(NodeType.RESOURCE, r -> r.publicId("urn:resource:1"));
 
         var response = testUtils.getResource("/v1/resources/urn:resource:1");
-        final var resource = testUtils.getObject(ResourceDTO.class, response);
+        final var resource = testUtils.getObject(NodeDTO.class, response);
 
         assertEquals("", resource.getPath());
     }
@@ -84,7 +87,7 @@ public class ResourcesTest extends RestTest {
                 .child(t -> t.nodeType(NodeType.TOPIC).resource(true, r -> r.name("Gas giants"))));
 
         var response = testUtils.getResource("/v1/resources");
-        final var resources = testUtils.getObject(ResourceDTO[].class, response);
+        final var resources = testUtils.getObject(NodeDTO[].class, response);
 
         assertEquals(2, resources.length);
         assertAnyTrue(resources, s -> "The inner planets".equals(s.getName()));
@@ -111,7 +114,7 @@ public class ResourcesTest extends RestTest {
 
         {
             final var response = testUtils.getResource("/v1/resources?contentURI=urn:test:1");
-            final var resources = testUtils.getObject(ResourceDTO[].class, response);
+            final var resources = testUtils.getObject(NodeDTO[].class, response);
 
             assertEquals(1, resources.length);
             assertEquals("The inner planets", resources[0].getName());
@@ -119,7 +122,7 @@ public class ResourcesTest extends RestTest {
 
         {
             final var response = testUtils.getResource("/v1/resources?contentURI=urn:test:2");
-            final var resources = testUtils.getObject(ResourceDTO[].class, response);
+            final var resources = testUtils.getObject(NodeDTO[].class, response);
 
             assertEquals(1, resources.length);
             assertEquals("Gas giants", resources[0].getName());
@@ -139,7 +142,7 @@ public class ResourcesTest extends RestTest {
 
         {
             final var response = testUtils.getResource("/v1/resources?isVisible=false");
-            final var resources = testUtils.getObject(ResourceDTO[].class, response);
+            final var resources = testUtils.getObject(NodeDTO[].class, response);
 
             assertEquals(2, resources.length);
             assertAnyTrue(resources, s -> "The inner planets".equals(s.getName()));
@@ -147,14 +150,14 @@ public class ResourcesTest extends RestTest {
         }
         {
             final var response = testUtils.getResource("/v1/resources?key=custom&value=value");
-            final var resources = testUtils.getObject(ResourceDTO[].class, response);
+            final var resources = testUtils.getObject(NodeDTO[].class, response);
 
             assertEquals(1, resources.length);
             assertEquals("Gas giants", resources[0].getName());
         }
         {
             final var response = testUtils.getResource("/v1/resources?value=value");
-            final var resources = testUtils.getObject(ResourceDTO[].class, response);
+            final var resources = testUtils.getObject(NodeDTO[].class, response);
 
             assertEquals(2, resources.length);
             assertAnyTrue(resources, s -> "Gas giants".equals(s.getName()));
@@ -162,7 +165,7 @@ public class ResourcesTest extends RestTest {
         }
         {
             final var response = testUtils.getResource("/v1/resources?key=key&isVisible=false");
-            final var resources = testUtils.getObject(ResourceDTO[].class, response);
+            final var resources = testUtils.getObject(NodeDTO[].class, response);
 
             assertEquals(1, resources.length);
             assertEquals("Observing Uranus", resources[0].getName());
@@ -342,7 +345,7 @@ public class ResourcesTest extends RestTest {
         URI id = resource.getPublicId();
 
         var response = testUtils.getResource("/v1/resources/" + id);
-        final var result = testUtils.getObject(ResourceDTO.class, response);
+        final var result = testUtils.getObject(NodeDTO.class, response);
 
         assertEquals(id, result.getId());
     }
@@ -396,15 +399,14 @@ public class ResourcesTest extends RestTest {
                 .publicId("urn:topic:1").contentUri(URI.create("urn:article:6662")).resource(resource, true));
 
         var response = testUtils.getResource("/v1/resources/" + resource.getPublicId() + "/full");
-        final var result = testUtils.getObject(ResourceWithParentsDTO.class, response);
+        final var result = testUtils.getObject(NodeWithParents.class, response);
 
         assertEquals(resource.getPublicId(), result.getId());
         assertEquals(resource.getName(), result.getName());
         assertEquals(1, result.getResourceTypes().size());
         assertEquals(resourceType.getName(), result.getResourceTypes().iterator().next().getName());
-        assertEquals(0, result.getFilters().size());
         assertEquals(1, result.getParents().size());
-        final NodeWithResourceConnectionDTO t = result.getParents().iterator().next();
+        final NodeChildDTO t = result.getParents().iterator().next();
         assertEquals(topic.getName(), t.getName());
         assertTrue(t.isPrimary());
         assertEquals(URI.create("urn:article:6662"), t.getContentUri());
@@ -415,7 +417,7 @@ public class ResourcesTest extends RestTest {
         testSeeder.resourceInDualSubjectsTestSetup();
 
         var response = testUtils.getResource("/v1/resources/urn:resource:1/full");
-        final var result = testUtils.getObject(ResourceWithParentsDTO.class, response);
+        final var result = testUtils.getObject(NodeWithParents.class, response);
         assertNotNull(result.getPaths());
         assertEquals(2, result.getPaths().size());
     }
@@ -425,7 +427,7 @@ public class ResourcesTest extends RestTest {
         testSeeder.resourceInDualSubjectsTestSetup();
 
         var response = testUtils.getResource("/v1/resources/urn:resource:1");
-        final var result = testUtils.getObject(ResourceWithParentsDTO.class, response);
+        final var result = testUtils.getObject(NodeWithParents.class, response);
         assertNotNull(result.getPaths());
         assertEquals(2, result.getPaths().size());
     }
@@ -434,7 +436,7 @@ public class ResourcesTest extends RestTest {
     public void can_get_resource_connection_id() throws Exception {
         Node topic = builder.node(t -> t.nodeType(NodeType.TOPIC).publicId("urn:topic:1").resource());
         var response = testUtils.getResource("/v1/topics/urn:topic:1/resources");
-        final var result = testUtils.getObject(ResourceChildDTO[].class, response);
+        final var result = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals(first(topic.getResourceChildren()).getPublicId(), result[0].getConnectionId());
     }
@@ -443,7 +445,7 @@ public class ResourcesTest extends RestTest {
     public void can_get_resource_connections_with_metadata() throws Exception {
         Node topic = builder.node(t -> t.nodeType(NodeType.TOPIC).publicId("urn:topic:1").resource());
         var response = testUtils.getResource("/v1/topics/urn:topic:1/resources");
-        final var result = testUtils.getObject(ResourceChildDTO[].class, response);
+        final var result = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals(first(topic.getResourceChildren()).getPublicId(), result[0].getConnectionId());
         assertAllTrue(result, connection -> connection.getMetadata() != null);
@@ -459,7 +461,7 @@ public class ResourcesTest extends RestTest {
                         st -> st.publicId("urn:topic:2").resource(r -> r.name("b").publicId("urn:resource:2"))));
 
         var response = testUtils.getResource("/v1/topics/urn:topic:1343/resources?recursive=true");
-        final var result = testUtils.getObject(ResourceChildDTO[].class, response);
+        final var result = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals(first(builder.node("topic").getResourceChildren()).getPublicId(), result[0].getConnectionId());
         assertEquals(first(builder.node("subtopic").getResourceChildren()).getPublicId(), result[1].getConnectionId());
@@ -479,7 +481,7 @@ public class ResourcesTest extends RestTest {
                                         .resource(true, r -> r.name("resource aab").contentUri("urn:article:aab"))))));
 
         var response = testUtils.getResource("/v1/topics/urn:topic:a/resources?recursive=true");
-        final var result = testUtils.getObject(ResourceChildDTO[].class, response);
+        final var result = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals(4, result.length);
         assertAnyTrue(result,
@@ -491,14 +493,14 @@ public class ResourcesTest extends RestTest {
         assertAnyTrue(result,
                 r -> "resource aab".equals(r.getName()) && "urn:article:aab".equals(r.getContentUri().toString()));
         assertAllTrue(result, r -> !r.getPaths().isEmpty());
-        assertAllTrue(result, ResourceChildDTO::isPrimary);
+        assertAllTrue(result, NodeChildDTO::isPrimary);
     }
 
     @Test
     public void resources_by_topic_id_recursively_are_ordered_by_rank_in_parent() throws Exception {
         testSeeder.resourcesBySubjectIdTestSetup();
         var response = testUtils.getResource("/v1/topics/urn:topic:5/resources?recursive=true");
-        final var result = testUtils.getObject(ResourceChildDTO[].class, response);
+        final var result = testUtils.getObject(NodeChildDTO[].class, response);
         assertEquals(6, result.length);
         assertEquals("urn:resource:3", result[0].getId().toString());
         assertEquals("urn:resource:5", result[1].getId().toString());
@@ -528,13 +530,13 @@ public class ResourcesTest extends RestTest {
 
         {
             var response = testUtils.getResource("/v1/topics/urn:topic:rt:1201/resources");
-            final var result = testUtils.getObject(ResourceChildDTO[].class, response);
+            final var result = testUtils.getObject(NodeChildDTO[].class, response);
             assertEquals(1, result.length);
             assertTrue(result[0].isPrimary());
         }
         {
             var response = testUtils.getResource("/v1/topics/urn:topic:rt:1202/resources");
-            final var result = testUtils.getObject(ResourceChildDTO[].class, response);
+            final var result = testUtils.getObject(NodeChildDTO[].class, response);
             assertEquals(1, result.length);
             assertFalse(result[0].isPrimary());
         }
@@ -553,14 +555,14 @@ public class ResourcesTest extends RestTest {
                                         r -> r.publicId("urn:resource:aab"))))));
 
         var response = testUtils.getResource("/v1/topics/urn:topic:a/resources?recursive=true");
-        final var result = testUtils.getObject(ResourceChildDTO[].class, response);
+        final var result = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals(4, result.length);
         assertAnyTrue(result, r -> "/subject:1/topic:a/resource:a".equals(r.getPath()));
         assertAnyTrue(result, r -> "/subject:1/topic:a/topic:aa/resource:aa".equals(r.getPath()));
         assertAnyTrue(result, r -> "/subject:1/topic:a/topic:aa/topic:aaa/resource:aaa".equals(r.getPath()));
         assertAnyTrue(result, r -> "/subject:1/topic:a/topic:aa/topic:aab/resource:aab".equals(r.getPath()));
-        assertAllTrue(result, ResourceChildDTO::isPrimary);
+        assertAllTrue(result, NodeChildDTO::isPrimary);
     }
 
     @Test
@@ -571,7 +573,7 @@ public class ResourcesTest extends RestTest {
                         .resource(r -> r.name("resource 1")).resource(r -> r.name("resource 2"))));
 
         var response = testUtils.getResource("/v1/topics/urn:topic:1/resources");
-        final var result = testUtils.getObject(ResourceChildDTO[].class, response);
+        final var result = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals(2, result.length);
         assertAnyTrue(result, r -> "resource 1".equals(r.getName()));
@@ -583,11 +585,11 @@ public class ResourcesTest extends RestTest {
         testSeeder.resourceWithRelevancesTestSetup();
 
         var response = testUtils.getResource("/v1/topics/urn:topic:1/resources?relevance=urn:relevance:core");
-        final var resources = testUtils.getObject(ResourceChildDTO[].class, response);
+        final var resources = testUtils.getObject(NodeChildDTO[].class, response);
         assertEquals(9, resources.length);
 
         var response2 = testUtils.getResource("/v1/topics/urn:topic:1/resources?relevance=urn:relevance:supplementary");
-        final var resources2 = testUtils.getObject(ResourceChildDTO[].class, response2);
+        final var resources2 = testUtils.getObject(NodeChildDTO[].class, response2);
         assertEquals(1, resources2.length);
 
     }
@@ -598,25 +600,25 @@ public class ResourcesTest extends RestTest {
 
         {
             var response = testUtils.getResource("/v1/topics/urn:topic:1:1/resources?relevance=urn:relevance:core");
-            final var resources = testUtils.getObject(ResourceChildDTO[].class, response);
+            final var resources = testUtils.getObject(NodeChildDTO[].class, response);
             assertEquals(5, resources.length);
         }
         {
             var response = testUtils.getResource("/v1/topics/urn:topic:2:1/resources?relevance=urn:relevance:core");
-            final var resources = testUtils.getObject(ResourceChildDTO[].class, response);
+            final var resources = testUtils.getObject(NodeChildDTO[].class, response);
             assertEquals(5, resources.length);
         }
 
         {
             var response = testUtils
                     .getResource("/v1/topics/urn:topic:1:1/resources?relevance=urn:relevance:supplementary");
-            final var resources = testUtils.getObject(ResourceChildDTO[].class, response);
+            final var resources = testUtils.getObject(NodeChildDTO[].class, response);
             assertEquals(0, resources.length);
         }
         {
             MockHttpServletResponse response = testUtils
                     .getResource("/v1/topics/urn:topic:2:1/resources?relevance=urn:relevance:supplementary");
-            final var resources = testUtils.getObject(ResourceChildDTO[].class, response);
+            final var resources = testUtils.getObject(NodeChildDTO[].class, response);
             assertEquals(5, resources.length);
         }
     }
@@ -627,25 +629,25 @@ public class ResourcesTest extends RestTest {
 
         {
             var response = testUtils.getResource("/v1/topics/urn:topic:1:1/resources?relevance=urn:relevance:core");
-            final var resources = testUtils.getObject(ResourceChildDTO[].class, response);
+            final var resources = testUtils.getObject(NodeChildDTO[].class, response);
             assertEquals(5, resources.length);
         }
         {
             var response = testUtils.getResource("/v1/topics/urn:topic:2:1/resources?relevance=urn:relevance:core");
-            final var resources = testUtils.getObject(ResourceChildDTO[].class, response);
+            final var resources = testUtils.getObject(NodeChildDTO[].class, response);
             assertEquals(5, resources.length);
         }
 
         {
             var response = testUtils
                     .getResource("/v1/topics/urn:topic:1:1/resources?relevance=urn:relevance:supplementary");
-            final var resources = testUtils.getObject(ResourceChildDTO[].class, response);
+            final var resources = testUtils.getObject(NodeChildDTO[].class, response);
             assertEquals(0, resources.length);
         }
         {
             var response = testUtils
                     .getResource("/v1/topics/urn:topic:2:1/resources?relevance=urn:relevance:supplementary");
-            final var resources = testUtils.getObject(ResourceChildDTO[].class, response);
+            final var resources = testUtils.getObject(NodeChildDTO[].class, response);
             assertEquals(5, resources.length);
         }
     }
@@ -655,7 +657,7 @@ public class ResourcesTest extends RestTest {
         testSeeder.resourcesBySubjectNodeIdTestSetup();
 
         var response = testUtils.getResource("/v1/subjects/urn:subject:1/resources");
-        final var resources = testUtils.getObject(ResourceDTO[].class, response);
+        final var resources = testUtils.getObject(NodeDTO[].class, response);
 
         assertEquals(10, resources.length);
         assertEquals("R:9", resources[0].getName());
@@ -679,7 +681,7 @@ public class ResourcesTest extends RestTest {
                 .getPublicId();
 
         var response = testUtils.getResource("/v1/subjects/" + id + "/resources");
-        final var resources = testUtils.getObject(ResourceChildDTO[].class, response);
+        final var resources = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals("urn:article:1", resources[0].getContentUri().toString());
         // assertEquals("/subject:1/topic:1:1/resource:1", resources[0].getPath());
@@ -700,7 +702,7 @@ public class ResourcesTest extends RestTest {
                 .getPublicId();
 
         var response = testUtils.getResource("/v1/subjects/" + id + "/resources");
-        final var resources = testUtils.getObject(ResourceChildDTO[].class, response);
+        final var resources = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals(3, resources.length);
 
@@ -728,7 +730,7 @@ public class ResourcesTest extends RestTest {
                                         .resourceType(rt -> rt.name("lecture"))))));
 
         var response = testUtils.getResource("/v1/subjects/urn:subject:1/resources");
-        final var resources = testUtils.getObject(ResourceChildDTO[].class, response);
+        final var resources = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals(3, resources.length);
         assertAnyTrue(resources, r -> r.getPath().equals("/subject:1/topic:1/resource:1"));
