+++ src/test/java/no/ndla/taxonomy/rest/v1/SubjectResourceTypesTest.java
@@ -8,7 +8,7 @@
 package no.ndla.taxonomy.rest.v1;
 
 import no.ndla.taxonomy.domain.NodeType;
-import no.ndla.taxonomy.service.dtos.ResourceChildDTO;
+import no.ndla.taxonomy.service.dtos.NodeChildDTO;
 import org.junit.jupiter.api.Test;
 
 import java.net.URI;
@@ -30,7 +30,7 @@ public class SubjectResourceTypesTest extends RestTest {
                 .getPublicId();
 
         var response = testUtils.getResource("/v1/subjects/" + id + "/resources");
-        var result = testUtils.getObject(ResourceChildDTO[].class, response);
+        var result = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals(2, result.length);
         assertEquals(2, result[0].getResourceTypes().size());
@@ -45,7 +45,7 @@ public class SubjectResourceTypesTest extends RestTest {
                 .getPublicId();
 
         var response = testUtils.getResource("/v1/subjects/" + id + "/resources");
-        var result = testUtils.getObject(ResourceChildDTO[].class, response);
+        var result = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals(1, result.length);
         assertEquals(0, result[0].getResourceTypes().size());
@@ -62,7 +62,7 @@ public class SubjectResourceTypesTest extends RestTest {
                 .resource(r -> r.name("a lecture").resourceType("lecture")))).getPublicId();
 
         var response = testUtils.getResource("/v1/subjects/" + id + "/resources?type=" + lecture);
-        var result = testUtils.getObject(ResourceChildDTO[].class, response);
+        var result = testUtils.getObject(NodeChildDTO[].class, response);
 
         assertEquals(2, result.length);
         assertAnyTrue(result, r -> "a lecture".equals(r.getName()));
