+++ src/test/java/no/ndla/taxonomy/rest/v1/TopicsTest.java
@@ -12,8 +12,7 @@ import no.ndla.taxonomy.domain.Node;
 import no.ndla.taxonomy.domain.NodeType;
 import no.ndla.taxonomy.rest.v1.commands.TopicCommand;
 import no.ndla.taxonomy.service.dtos.ConnectionIndexDTO;
-import no.ndla.taxonomy.service.dtos.MetadataDto;
-import no.ndla.taxonomy.service.dtos.NodeDTO;
+import no.ndla.taxonomy.service.dtos.NodeChildDTO;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,7 +24,6 @@ import java.util.Set;
 
 import static no.ndla.taxonomy.TestUtils.*;
 import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
@@ -47,7 +45,7 @@ public class TopicsTest extends RestTest {
                 .nodeType(NodeType.TOPIC).name("trigonometry").contentUri("urn:article:1").publicId("urn:topic:1")));
 
         MockHttpServletResponse response = testUtils.getResource("/v1/topics/urn:topic:1");
-        final var topic = testUtils.getObject(NodeDTO.class, response);
+        final var topic = testUtils.getObject(NodeChildDTO.class, response);
 
         assertEquals("trigonometry", topic.getName());
         assertEquals("urn:article:1", topic.getContentUri().toString());
@@ -63,7 +61,7 @@ public class TopicsTest extends RestTest {
         builder.node(NodeType.TOPIC, t -> t.publicId("urn:topic:1"));
 
         MockHttpServletResponse response = testUtils.getResource("/v1/topics/urn:topic:1");
-        final var topic = testUtils.getObject(NodeDTO.class, response);
+        final var topic = testUtils.getObject(NodeChildDTO.class, response);
 
         assertEquals("", topic.getPath());
     }
@@ -81,14 +79,14 @@ public class TopicsTest extends RestTest {
 
         {
             final var response = testUtils.getResource("/v1/topics?contentURI=urn:test:1");
-            final var topics = testUtils.getObject(NodeDTO[].class, response);
+            final var topics = testUtils.getObject(NodeChildDTO[].class, response);
             assertEquals(1, topics.length);
             assertEquals("photo synthesis", topics[0].getName());
         }
 
         {
             final var response = testUtils.getResource("/v1/topics?contentURI=urn:test:2");
-            final var topics = testUtils.getObject(NodeDTO[].class, response);
+            final var topics = testUtils.getObject(NodeChildDTO[].class, response);
             assertEquals(1, topics.length);
             assertEquals("trigonometry", topics[0].getName());
         }
@@ -112,7 +110,7 @@ public class TopicsTest extends RestTest {
 
         {
             final var response = testUtils.getResource("/v1/topics?value=value");
-            final var nodes = testUtils.getObject(NodeDTO[].class, response);
+            final var nodes = testUtils.getObject(NodeChildDTO[].class, response);
             assertEquals(1, nodes.length);
             assertEquals("photo synthesis", nodes[0].getName());
             assertNotNull(nodes[0].getMetadata());
@@ -121,7 +119,7 @@ public class TopicsTest extends RestTest {
         }
         {
             final var response = testUtils.getResource("/v1/topics?key=test&value=value");
-            final var nodes = testUtils.getObject(NodeDTO[].class, response);
+            final var nodes = testUtils.getObject(NodeChildDTO[].class, response);
             assertEquals(1, nodes.length);
             assertEquals("photo synthesis", nodes[0].getName());
             assertNotNull(nodes[0].getMetadata());
@@ -130,7 +128,7 @@ public class TopicsTest extends RestTest {
         }
         {
             final var response = testUtils.getResource("/v1/topics?key=test&value=value2");
-            final var nodes = testUtils.getObject(NodeDTO[].class, response);
+            final var nodes = testUtils.getObject(NodeChildDTO[].class, response);
             assertEquals(1, nodes.length);
             assertEquals("trigonometry", nodes[0].getName());
             assertNotNull(nodes[0].getMetadata());
@@ -147,7 +145,7 @@ public class TopicsTest extends RestTest {
                 s -> s.isContext(true).name("Maths").child(NodeType.TOPIC, t -> t.name("trigonometry")));
 
         MockHttpServletResponse response = testUtils.getResource("/v1/topics");
-        final var topics = testUtils.getObject(NodeDTO[].class, response);
+        final var topics = testUtils.getObject(NodeChildDTO[].class, response);
         assertEquals(2, topics.length);
 
         assertAnyTrue(topics, t -> "photo synthesis".equals(t.getName()));
@@ -200,7 +198,7 @@ public class TopicsTest extends RestTest {
         testSeeder.subtopicsByNodeIdAndRelevanceTestSetup();
 
         MockHttpServletResponse response = testUtils.getResource("/v1/topics/urn:topic:1/topics");
-        final var subtopics = testUtils.getObject(NodeDTO[].class, response);
+        final var subtopics = testUtils.getObject(NodeChildDTO[].class, response);
         assertEquals(7, subtopics.length);
 
         assertEquals("urn:topic:2", subtopics[0].getId().toString());
@@ -216,7 +214,7 @@ public class TopicsTest extends RestTest {
         testSeeder.subtopicsByNodeIdAndRelevanceTestSetup();
 
         MockHttpServletResponse response = testUtils.getResource("/v1/topics/urn:topic:1/topics");
-        final var subtopics = testUtils.getObject(NodeDTO[].class, response);
+        final var subtopics = testUtils.getObject(NodeChildDTO[].class, response);
         assertEquals(7, subtopics.length, "Unfiltered subtopics");
 
         assertAllTrue(subtopics, subtopic -> subtopic.getMetadata() != null);
